Add bundled schema output (#83)
There are [various tools](https://json-schema.org/tools?query=&sortBy=name&sortOrder=ascending&groupBy=toolingTypes&licenses=&languages=&drafts=&toolingTypes=&environments=&showObsolete=false#bundler) to do this in a post-processing step. However, the output from the tools I tried failed to work with the IDL plugins and other JSON schema libraries I tried.

This PR updates the plugin to always output a bundled representation of the schema as a '*.schema.bundle.json' file that works with the IDL plugins and the JSON schema libraries I tested.

Fixes: https://github.com/bufbuild/protoschema-plugins/issues/20

As the amount of output was getting large, so also added a 'target' option to limit the output.